Show filter result and record count (Closes #14)
Bottom of Asset table listing now show the index of the current record
set being displayed, along with the total record count.

For example, when viewing the full asset table the first page should
read "Showing 1-25 of 121 results".  When navigating to page 2 is should
read "Showing 26-50 of 121 results.

If a user filters the result set the 'total row count' returned will be
the total number of records in the filtered set, not the total number of
records in the table.