Work around autocomplete bug when new tag created (closes #57)
The select presents methods to create a new tag using the built in
select2 method, and a method from dal or taggit [not sure].  If use the
latter the new tag was repeated 4 times in the select box. Disable new
tag addition in the autcomplete url.

New tags still appear to be created